+++ src/main/php/lang/reflection/Annotation.class.php
@@ -11,19 +11,36 @@ class Annotation implements Value {
     $this->arguments= $arguments;
   }
 
-  public function type() { return $this->type; }
+  public function type(): string { return $this->type; }
 
-  public function name() {
+  /** Returns lowercase name */
+  public function name(): string {
     return strtolower(false === ($p= strrpos($this->type, '\\'))
       ? $this->type
       : substr($this->type, $p + 1
     ));
   }
 
-  public function arguments() { return $this->arguments; }
+  /** Returns all arguments */
+  public function arguments(): array { return $this->arguments; }
 
+  /**
+   * Returns a given argument, or NULL if it doesn't exist
+   *
+   * @param  string|int $key
+   * @return var
+   */
   public function argument($key) { return $this->arguments[$key] ?? null; }
 
+  /** @return object */
+  public function newInstance() { return new $this->type(...$this->arguments); }
+
+  /**
+   * Checks whether this annotation is of a given type
+   *
+   * @param  lang.XPClass|lang.reflection.Type $type
+   * @return bool
+   */
   public function is($type) {
     if ($type instanceof Type || $type instanceof XPClass) {
       $compare= $type->literal();